removed props
