+++ app/infrastructure/repositories/artistry/сompositions.repository.ts
@@ -24,6 +24,7 @@ export const compositionsRepository = {
       .populate('genres')
       .populate({ path: 'opusId', model: Opus })
       .lean();
+    console.log(compositions);
     if (!compositions || compositions.length === 0) {
       return [];
     }
